➕
Type(s): ADD

+++ app/src/main/java/com/george/freenowassessment/repositories/VehicleDataManager.kt
@@ -5,6 +5,9 @@ import com.george.freenowassessment.data.local.Vehicle
 import com.george.freenowassessment.data.local.VehicleDao
 import com.george.freenowassessment.data.local.VehicleList
 import com.george.freenowassessment.data.remote.responses.VehicleData
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 /**
@@ -21,23 +24,30 @@ class VehicleDataManager(
     * this function update all local data from db with [VehicleData] from API
     * */
     suspend fun storeData(bound: String) {
-        val vehicles = VehicleList(dao.allVehicles(bound) as ArrayList<Vehicle>)
-        vehicles.deactivateAll()
-
-        list.forEach { data ->
-            if (vehicles.hasVehicle(data.id)) {
-                val vehicle = vehicles.getVehicle(data.id)
-                if(vehicle?.equals(data) == false) {
-                    vehicles.update(data, geocoder)
-                } else {
-                    vehicle?.state = data.state
+        GlobalScope.launch(Dispatchers.IO) {
+            val allVehicles = dao.allVehicles(bound)
+            val vehicles = VehicleList()
+            if (allVehicles.isNotEmpty()) {
+                vehicles.add(allVehicles)
+                vehicles.deactivateAll()
+                list.forEach { data ->
+                    if (vehicles.hasVehicle(data.id)) {
+                        val vehicle = vehicles.getVehicle(data.id)
+                        if (vehicle?.equals(data) == false) {
+                            vehicles.update(data, geocoder)
+                        } else {
+                            vehicle?.state = data.state
+                        }
+                    } else {
+                        vehicles.add(data, geocoder, bound)
+                    }
                 }
+                dao.deleteAll(bound)
             } else {
-                vehicles.add(data, geocoder, bound)
+                vehicles.add(list, geocoder, bound)
             }
+            dao.insertAll(vehicles.vehicles)
         }
-        dao.deleteAll(bound)
-        dao.insertAll(vehicles.vehicles)
     }
 
 }
